+++ .env
@@ -2,4 +2,4 @@ UPDATE_REPORT=true
 MILO=true
 SELF_MANAGED_AUTOBAHN=false
 CASES=1.*,2.*
-NRDP=pvm current run --
+NRDP=pvm current run
