+++ autobahn/runner/index.ts
@@ -7,7 +7,6 @@ import shell from 'shelljs';
 import { Platform } from '../../dist/milo.node';
 import { systemReq } from './sys-requirements';
 import { root } from './paths';
-import { start, stop } from './docker';
 import { runAutobahnTests, AutobahnOpts } from './run-test';
 
 export default async function autobahn(WebSocketClass: WebSocket, opts: AutobahnOpts): Promise<number> {
